Move character struct to lib/characters/character.ex
